test(lowering): rename DifferentialSpec to RealIrCodecRoundTripSpec, correct its claims
"Differential" was a misnomer, and the docstring overstated what the
test proves: it claimed that disagreement between the two sides means
"either the lowering drops information or the derived codecs are not
round-trip stable", but the assertion is decode(encode(lowered)) ==
lowered - both sides derive from the same already-lowered value, so it
cannot detect information the lowering itself drops on the way from v3.
It also claimed the fixture was a "JVM-only test resource"; Mill's
resources task is Task.Sources("resources") and isn't platform-remapped
(verified: morphir.model.lowering.jvm.test.resources and
...js.test.resources resolve to the same directory), so the fixture
sits on the JS classpath too - only test/src-jvm is genuinely JVM-only.

Rewrite the docstring to state what the test actually proves: the
lowering is total and non-throwing over 745 KB of real morphir-elm
output, and the derived codecs round-trip shapes real IR actually
produces. Add a second test giving a genuine (if coarse) v3-to-lowered
correspondence check - matching module/type/value counts between the v3
source and the lowered tree - since the round-trip alone proves nothing
about the two sides agreeing with each other.
